refactor: new header component file struxure
